+++ locales/fr/main.lang.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * @package Dotclear
+ *
+ * @copyright Olivier Meunier & Association Dotclear
+ * @copyright GPL-2.0-only
+ */
+#
+# DOT NOT MODIFY THIS FILE !
+#
+
+l10n::$locales['Mail to log']                    = 'Mail en log';
+l10n::$locales['Do not send mails but log them'] = 'Ne pas envoyer les mails mais les loguer';